fix(OUT-2104): finalized virtualization of tasks list on list and board.
